requirements.txt
django
psycopg2-binary
pillow
djangorestframework
django-filter
djangorestframework-simplejwt
coverage
drf-yasg
django-cors-headers
requests
stripe
python-dotenv
forex-python